Moved CSS Files to Folder & Asset Logo To Assets
Moved CSS and & Assets into the respective folders each (excluding defaults from default install i.e. logo.svg
