+++ frontend/src/DashboardPage.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import './DashboardPage.css';
+import './css/DashboardPage.css';
 import logo from './assets/FinalLogo.png';
 import MainContent from './MainContent';
