--- frontend/src/LoginForm.css
@@ -1,122 +1,0 @@
-.login-form-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 20px;
-  background-color: #111920;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  width: 90%;
-  max-width: 400px;
-  max-height: 90vh;
-  overflow: auto;
-  color: white;
-}
-
-.login-logo {
-  max-height: 80px;
-  width: auto;
-  margin-bottom: 24px;
-}
-
-.login-content {
-  width: 100%;
-}
-
-.login-input {
-  padding: 12px;
-  margin-bottom: 16px;
-  border: 2px solid #2a2f33;
-  background-color: #fff;
-  color: black;
-  border-radius: 20px;
-  width: calc(100% - 24px);
-  transition: border-color 0.3s;
-}
-
-.login-input:focus {
-  outline: none;
-  border-color: #f44336;
-}
-
-.login-submit-btn {
-  padding: 12px;
-  margin-bottom: 16px;
-  border: none;
-  border-radius: 20px;
-  background-image: linear-gradient(to right, #2196F3, #f44336);
-  color: white;
-  width: calc(100% - 24px);
-  font-weight: bold;
-  cursor: pointer;
-  transition: background-image 0.3s;
-}
-
-.login-submit-btn:hover {
-  background-image: linear-gradient(to right, #1976D2, #d32f2f);
-}
-
-.forgot-password-link, .signin-username-link {
-  display: block;
-  text-align: center;
-  color: #2196F3;
-  margin: 16px 0;
-  text-decoration: none;
-  transition: text-decoration 0.3s;
-}
-
-.forgot-password-link:hover, .signin-username-link:hover {
-  text-decoration: underline;
-}
-
-.login-form-footer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 16px;
-}
-
-.close-button, .signup-button {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1em;
-  border: 2px solid transparent; /* Add a transparent border for consistent sizing */
-  padding: 12px;
-  border-radius: 20px;
-  color: white;
-  cursor: pointer;
-  font-weight: bold;
-  margin: 4px;
-  transition: background-color 0.3s;
-  width: 130px;
-  height: 40px;
-  box-sizing: border-box;
-  text-decoration: none;
-}
-
-.close-button {
-  background-color: #f44336;
-}
-
-.signup-button {
-  background-color: #2196F3;
-}
-
-.close-button:hover, .signup-button:hover {
-  opacity: 0.9;
-}
-
-.forgot-password-link {
-  display: block;
-  text-align: center;
-  color: #2196F3;
-  margin: 16px 0;
-  text-decoration: none;
-  transition: text-decoration 0.3s;
-}
-
-.forgot-password-link:hover {
-  text-decoration: underline;
-}
