+++ frontend/src/LoginForm.js
@@ -1,6 +1,6 @@
 import { useParams, useNavigate } from 'react-router-dom';
 import React, { useState } from 'react';
-import './LoginForm.css';
+import './css/LoginForm.css';
 import logo from './assets/FinalLogo.png';
 import { instance } from './App';
