+++ frontend/src/css/AnimeCard.css
