+++ frontend/src/css/App.css
@@ -1,0 +1,111 @@
+@import url('https://fonts.googleapis.com/css2?family=Bangers&display=swap');
+
+#background-video {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  min-width: 100%;
+  min-height: 100%;
+  width: auto;
+  height: auto;
+  z-index: -1;
+}
+
+.video-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  z-index: 1;
+}
+
+.topbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 80px;
+  background-color: #111920;
+  z-index: 2;
+  padding: 0 20px;
+}
+
+.topbar-logo {
+  height: 60px;
+  width: auto;
+}
+
+.buttons-container {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+.topbar-btn, .get-started-btn {
+  background: none;
+  color: white;
+  border: 2px solid transparent;
+  padding: 15px 30px;
+  font-size: 18px;
+  cursor: pointer;
+  margin: 0 10px;
+  font-family: 'Bangers', cursive;
+  text-decoration: none;
+  transition: all 0.3s;
+  border-radius: 5px;
+}
+
+.get-started-btn {
+  border: 2px solid white;
+}
+
+.topbar-btn:hover, .get-started-btn:hover {
+  background-color: white;
+  color: #111920;
+  border-color: transparent;
+}
+
+.App {
+  text-align: center;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
