+++ frontend/src/css/DashboardPage.css
@@ -1,0 +1,62 @@
+
+.App {
+  text-align: center;
+}
+
+.dashboard {
+  background-color: #2a2f36;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+}
+
+.topbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 80px;
+  background-color: #111920;
+  z-index: 2;
+  padding: 0 20px;
+  box-shadow: 0 2px 4px rgba(0,0,0,.5);
+}
+
+.topbar-logo {
+  height: 60px;
+  width: auto;
+}
+
+.topbar-btn, .log-out-btn {
+  background: none;
+  color: white;
+  border: 2px solid transparent;
+  padding: 10px 20px;
+  font-size: 16px;
+  cursor: pointer;
+  margin: 0 10px;
+  font-family: 'Bangers', cursive;
+  transition: all 0.3s;
+  border-radius: 5px;
+  text-decoration: none;
+}
+
+.log-out-btn {
+  border: 2px solid white;
+}
+
+.topbar-btn:hover, .log-out-btn:hover {
+  background-color: white;
+  color: #111920;
+  border-color: transparent;
+}
+
+body {
+  padding-top: 80px;
+}
+
